+++ CHANGELOG.md
@@ -64,3 +64,4 @@
 -   Overview command.
 -   String linting rules (yet have to increase level of linting).
 -   STW resources command.
+-   Image utils